LPD-84025 Consolidate locked query detection logic in BaseDB
